+++ Source/Libraries/Common/Common.csproj
@@ -56,6 +56,7 @@
     </Compile>
     <Compile Include="Logging.cs" />
     <Compile Include="Extensions\RandomExtensions.cs" />
+    <Compile Include="Extensions\StringExtensions.cs" />
     <Compile Include="Objects\OperationResults.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="StaticTools.cs" />
